Stage 1b of workout→session rename: retarget torii MCP tools + coach logging to workoutlogs
The torii MCP workout tools and the coach's log_workout wrote to the
dead `workouts` collection, so sessions logged via MCP or chat never
appeared in the app (the app reads workoutlogs). This retargets both
paths at the real logging collection and only then deletes the legacy
model — workoutTools.js required its controller at module load, so
deletion had to follow the retarget.

- Extract routes/workoutLogs.js inline CRUD into
  controllers/workoutLogController.js (routes now delegate); includes
  exercise-id resolution, calendar-event creation + backlink, cascade
  delete
- Retarget all six torii tools (list/get/create/update/delete_workout,
  get_workout_stats) at the log controller with log-shaped schemas
  (startedAt/completedAt/durationMinutes; no status enum — logs are
  performed sessions)
- Coach log_workout now inserts a WorkoutLog-shaped doc into
  workoutlogs and mirrors the backend path by creating the completed
  calendar event + backlink; get_workout_history reads workoutlogs
  (startedAt/actualDuration)
- Delete models/Workout.js, controllers/workoutController.js, and the
  Workout-shaped validateWorkout from middleware/validation.js
